+++ Geode/Types/VarTypeSpecifier.cs
@@ -1,4 +1,5 @@
-﻿using Datapack.Net.Utils;
+﻿using Datapack.Net.Data;
+using Datapack.Net.Utils;
 using Geode.Errors;
 using Geode.Values;
 
@@ -7,6 +8,7 @@ namespace Geode.Types
 	public class VarTypeSpecifier : TypeSpecifier
 	{
 		public override LiteralValue DefaultValue => throw new InvalidTypeError("var");
+		public override NBTType EffectiveType => NBTType.Compound;
 
 		public override NamespacedID ID => "amethyst:var";