+++ Geode/Values/MethodValue.cs
@@ -7,7 +7,7 @@ namespace Geode.Values
 		public readonly FunctionValue BaseFunction = val;
 		public readonly ValueRef Self = self;
 
-		public override void Call(RenderContext ctx, ValueRef[] args, bool applyGuard = true) => BaseFunction.Call(ctx, [Self, .. args], applyGuard);
+		public override void Call(RenderContext ctx, ValueRef[] args) => BaseFunction.Call(ctx, [Self, .. args]);
 
 		public override IFunctionLike CloneWithType(FunctionTypeSpecifier type) => new MethodValue(BaseFunction, Self, type);
 	}