+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* 
+* Allow constants in branches
+* Dynamic function calls